Scale browser test harness structure
